+++ packages/BetterReflectionDocBlock/src/Exception/NotImplementedYetException.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Symplify\BetterReflectionDocBlock\Exception;
+
+use Exception;
+
+final class NotImplementedYetException extends Exception
+{
+}
