+++ packages/CodingStandard/src/Fixer/Commenting/RemoveUselessDocBlockFixer.php
@@ -15,6 +15,7 @@ use PhpCsFixer\Tokenizer\Token;
 use PhpCsFixer\Tokenizer\Tokens;
 use PhpCsFixer\WhitespacesFixerConfig;
 use SplFileInfo;
+use Symplify\BetterReflectionDocBlock\PhpDocParser\TypeResolver;
 use Symplify\TokenRunner\DocBlock\DescriptionAnalyzer;
 use Symplify\TokenRunner\DocBlock\ParamAndReturnTagAnalyzer;
 use Symplify\TokenRunner\Wrapper\FixerWrapper\DocBlockWrapper;
@@ -48,17 +49,24 @@ final class RemoveUselessDocBlockFixer implements DefinedFixerInterface, Configu
      */
     private $methodWrapperFactory;
 
+    /**
+     * @var TypeResolver
+     */
+    private $typeResolver;
+
     public function __construct(
         DescriptionAnalyzer $descriptionAnalyzer,
         ParamAndReturnTagAnalyzer $paramAndReturnTagAnalyzer,
         MethodWrapperFactory $methodWrapperFactory,
-        WhitespacesFixerConfig $whitespacesFixerConfig
+        WhitespacesFixerConfig $whitespacesFixerConfig,
+        TypeResolver $typeResolver
     ) {
         $this->descriptionAnalyzer = $descriptionAnalyzer;
         $this->paramAndReturnTagAnalyzer = $paramAndReturnTagAnalyzer;
         $this->configure([]);
         $this->methodWrapperFactory = $methodWrapperFactory;
         $this->whitespacesFixerConfig = $whitespacesFixerConfig;
+        $this->typeResolver = $typeResolver;
     }
 
     public function getDefinition(): FixerDefinitionInterface
@@ -166,7 +174,7 @@ public function getCount(): int
             return;
         }
 
-        $docType = $docBlockWrapper->resolveDocType($returnTagValue->type);
+        $docType = $this->typeResolver->resolveDocType($returnTagValue->type);
 
         $returnTagDescription = $returnTagValue->description;
